requirements.txt
django==3.0.*
psycopg2-binary
markdown 
django-filter
pyjwt~=2.3.0
django-cors-headers
djangorestframework~=3.13.1
sentry-sdk